Convert spellusable modules back to JS to fix the crashes, until I figure out the dependency issues
