+++ src/parser/hunter/marksmanship/modules/core/SpellUsable.js
@@ -1,18 +1,18 @@
 import SPELLS from 'common/SPELLS';
 import CoreSpellUsable from 'parser/shared/modules/SpellUsable';
-import { CastEvent } from 'parser/core/Events';
 
 class SpellUsable extends CoreSpellUsable {
   static dependencies = {
     ...CoreSpellUsable.dependencies,
   };
 
-  lastPotentialTriggerForRapidFireReset: CastEvent | null = null;
+  lastPotentialTriggerForRapidFireReset = null;
   rapidFireResets = 0;
 
-  on_byPlayer_cast(event: CastEvent) {
-    super.on_byPlayer_cast?.(event);
-
+  on_byPlayer_cast(event) {
+    if (super.on_byPlayer_cast) {
+      super.on_byPlayer_cast(event);
+    }
     const spellId = event.ability.guid;
     if (this.selectedCombatant.hasTrait(SPELLS.SURGING_SHOTS.id)) {
       if (spellId === SPELLS.AIMED_SHOT.id) {
@@ -23,7 +23,7 @@ class SpellUsable extends CoreSpellUsable {
     }
   }
 
-  beginCooldown(spellId: number, cooldownTriggerEvent: Object) {
+  beginCooldown(spellId, cooldownTriggerEvent) {
     if (spellId === SPELLS.RAPID_FIRE.id && this.selectedCombatant.hasTrait(SPELLS.SURGING_SHOTS.id)) {
       if (this.isOnCooldown(spellId)) {
         this.rapidFireResets += 1;
